Refactor layout to use shared AppLayout and add new page routes
