+++ src/app/inventory/page.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+export default function Inventory() {
+  return (
+    <div className="inventory-page min-h-screen bg-background">
+      <div className="container mx-auto px-6 py-8">
+        <div className="page-header mb-8">
+          <h1 className="page-title text-3xl font-bold text-text">Inventory Management</h1>
+          <p className="page-description text-muted mt-2">
+            Manage your products, stock levels, and inventory operations
+          </p>
+        </div>
+
+        <div className="content-area">
+          <div className="placeholder-content bg-card border border-muted rounded-lg p-8 text-center">
+            <h2 className="text-xl font-semibold text-text mb-4">Inventory Page</h2>
+            <p className="text-muted">
+              This is the inventory management page. Add your inventory components here.
+            </p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
