+++ src/app/reports/layout.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { AppLayout } from '@/components/shared';
+
+export default function ReportsLayout({ children }: { children: React.ReactNode }) {
+  return <AppLayout>{children}</AppLayout>;
+}
