+++ src/app/settings/layout.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { AppLayout } from '@/components/shared';
+
+export default function SettingsLayout({ children }: { children: React.ReactNode }) {
+  return <AppLayout>{children}</AppLayout>;
+}
