+++ src/components/shared/index.ts
@@ -1,3 +1,4 @@
 export { Button } from './Button';
 export { Input } from './Input';
 export { Card, CardHeader, CardContent, CardFooter } from './Card';
+export { default as AppLayout } from './AppLayout';
